Fix: Moved Generate to a own folder
